removed getter from eventDate
