+++ .gitignore
@@ -42,4 +42,3 @@ docs
 .vscode
 **/.DS_Store
 web-server/public/content
-api-server/data
