+++ web-server/views/partials/aside.hbs
@@ -1,88 +1,13 @@
 <div class="col-xs-12 col-md-3 col-lg-3 col-xl-2 p-2">
 
   <nav class="nav flex-column sidebar text-center text-md-left">
-    <div class="semilist py-2">
-      <h5>Категорії</h5>
-      <form>
-        <div class="row">
-          <div class="col-xs-3 hidden-sm-up"></div>
-          <fieldset class="py-1 col-xs-6 col-sm-5 offset-sm-4 col-md-12 text-center text-sm-left ml-md-0">
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    Розробка ігор
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    Веб-розробка
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    3D Моделювання
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    Теорія алгоритмів
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    Розробка ігор
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    Веб-розробка
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    3D Моделювання
-                  </label>
-            </div>
-            <div class="form-check">
-              <label class="form-check-label">
-                    <input type="checkbox" class="form-check-input">
-                    Теорія алгоритмів
-                  </label>
-            </div>
-          </fieldset>
-        </div>
 
-        <button type="reset" class="btn btn-secondary">Онулити фільтри</button>
-      </form>
-    </div>
+    {{> categories}}
 
     <div class="dropdown-divider my-2"></div>
 
-    <div class="semilist py-2">
-      <h5>Архів</h5>
-
-      <a class="nav-link" href="#">Всі</a>
-
-      <a class="nav-link" data-toggle="collapse" href="#posts2017" aria-expanded="false" aria-controls="posts2017">2017 <i class="fa fa-caret-down" aria-hidden="true"></i></a>
-      <div class="collapse pl-3" id="posts2017">
-        <li class="nav-item">
-          <a class="nav-link" href="#">Подкаст</a>
-        </li>
-      </div>
-
-      <a class="nav-link" data-toggle="collapse" href="#posts2016" aria-expanded="false" aria-controls="posts2016">2016 <i class="fa fa-caret-down" aria-hidden="true"></i></a>
-      <div class="collapse pl-3" id="posts2016">
-        <li class="nav-item">
-          <a class="nav-link" href="#">Подкаст</a>
-        </li>
-      </div>
-    </div>
+    {{> archive}}
 
   </nav>
+
 </div>
