+++ web-server/views/partials/footer.hbs
@@ -1,7 +1,7 @@
 <footer class="px-5 py-3">
   <div class="container">
     <p class="mb-0 text-muted">
-      <i class="fa fa-copyright" aria-hidden="true"></i> 2017 Created by @vkhitev and @irmchk. Special for Belka Space.
+      <i class="fa fa-copyright" aria-hidden="true"></i> 2017 Created by <a href="https://github.com/vkhitev/">vkhitev</a>      and <a href="https://github.com/irmchk/">irmchk</a>
     </p>
   </div>
 </footer>
