+++ web-server/views/partials/navbar.hbs
@@ -1,26 +1,31 @@
 <nav class="navbar fixed-top navbar-toggleable-md navbar-inverse bg-orange">
   <div class="container mx-0 mx-sm-auto">
-    <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
-      aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
+
+    {{!-- Hamburger button, appears on small screens --}}
+    <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbar-content" aria-controls="navbar-content"
+      aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    
+    {{!-- Logo --}}
     <a class="navbar-brand" href="/">
       <img src="/img/logo.svg" lt="logo" class="belka-logo"> Belka
     </a>
 
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+    {{!-- Search --}}
+    <div class="collapse navbar-collapse" id="navbar-content">
       <ul class="navbar-nav ml-auto">
         <li class="nav-item">
           <div class="form-inline my-2 my-lg-0 d-flex">
-            <input class="form-control mr-sm-2" type="text" placeholder="Назва, автор, тег" id="search-input">
+            <input class="form-control mr-sm-2" type="text" placeholder="Название, автор" id="search-input">
             <button class="btn btn-outline-secondary my-2 my-sm-0" type="button" id="search-button">
-              Пошук 
+              Поиск 
               <i class="fa fa-search fa-lg" aria-hidden="true"></i>
             </button>
           </div>
         </li>
       </ul>
-
     </div>
+
   </div>
 </nav>
